+++ packages/api-reference/src/components/Content/Introduction/ClientSelector.vue
@@ -3,6 +3,7 @@ import { useMediaQuery } from '@vueuse/core'
 import { type TargetId, availableTargets } from 'httpsnippet-lite'
 import { ref } from 'vue'
 
+import { filterClients } from '../../../helpers'
 import { type SelectedClient, useTemplateStore } from '../../../stores/template'
 import { Icon } from '../../Icon'
 
@@ -46,7 +47,7 @@ const featuredClients = ref<SelectedClient[]>(
         },
         {
           targetKey: 'node',
-          clientKey: 'fetch',
+          clientKey: 'axios',
         },
         {
           targetKey: 'python',
@@ -65,7 +66,7 @@ const featuredClients = ref<SelectedClient[]>(
         },
         {
           targetKey: 'node',
-          clientKey: 'fetch',
+          clientKey: 'axios',
         },
         {
           targetKey: 'php',
@@ -153,7 +154,7 @@ function checkIfClientIsFeatured(client: SelectedClient) {
           :key="target.key"
           :label="target.title">
           <option
-            v-for="client in target.clients"
+            v-for="client in target.clients.filter(filterClients)"
             :key="client.key"
             :value="
               JSON.stringify({