+++ packages/api-reference/src/components/Content/ReferenceEndpoint/EndpointDetails.vue
@@ -1,0 +1,48 @@
+<script setup lang="ts">
+import { useOperation, useResponses } from '../../../hooks'
+import type { TransformedOperation } from '../../../types'
+import MarkdownRenderer from '../MarkdownRenderer.vue'
+import Parameters from './Parameters.vue'
+import RequestBody from './RequestBody.vue'
+
+const props = defineProps<{ operation: TransformedOperation }>()
+
+const { parameterMap } = useOperation(props)
+
+const { responses } = useResponses(props.operation)
+</script>
+<template>
+  <div class="endpoint-details">
+    <div class="endpoint-description">
+      <MarkdownRenderer :value="operation.description" />
+    </div>
+    <Parameters :parameters="parameterMap.path">
+      <template #title>Path Parameters</template>
+    </Parameters>
+    <Parameters :parameters="parameterMap.query">
+      <template #title>Query Parameters</template>
+    </Parameters>
+    <Parameters :parameters="parameterMap.header">
+      <template #title>Headers</template>
+    </Parameters>
+    <RequestBody :requestBody="operation.information?.requestBody">
+      <template #title>Body</template>
+    </RequestBody>
+    <Parameters :parameters="responses">
+      <template #title>Responses</template>
+    </Parameters>
+  </div>
+</template>
+<style scoped>
+.endpoint-details :deep(.endpoint-title) {
+  font-size: var(--theme-heading-4, var(--default-theme-heading-4));
+  font-weight: var(--theme-semibold, var(--default-theme-semibold));
+  color: var(--theme-color-1, var(--default-theme-color-1));
+  line-height: 1.45;
+  margin-top: 12px;
+  margin-bottom: 12px;
+}
+.endpoint-description {
+  margin-bottom: 24px;
+}
+</style>