+++ packages/api-reference/src/components/Content/ReferenceEndpoint/ExampleResponses/ExampleResponses.vue
@@ -16,7 +16,7 @@ import {
   CardTab,
   CardTabHeader,
 } from '../../../Card'
-import { Icon } from '../../../Icon'
+import { FlowIcon } from '../../../Icon'
 import MarkdownRenderer from '../../MarkdownRenderer.vue'
 // import Headers from './Headers.vue'
 import SelectExample from './SelectExample.vue'
@@ -89,8 +89,8 @@ const mergeAllObjects = (items: Record<any, any>[]): any => {
           class="code-copy"
           type="button"
           @click="() => copyToClipboard(currentJsonResponse?.example)">
-          <Icon
-            src="solid/interface-copy-clipboard"
+          <FlowIcon
+            icon="Clipboard"
             width="10px" />
         </button>
       </template>
@@ -306,6 +306,7 @@ const mergeAllObjects = (items: Record<any, any>[]): any => {
   margin: 6px;
 }
 .scalar-card-container {
+  flex: 1;
   background: var(--theme-background-2, var(--default-theme-background-2));
 }
 .scalar-card-container :deep(.cm-scroller) {