+++ packages/api-reference/src/components/Content/ReferenceEndpoint/ReferenceEndpoint.vue
@@ -10,9 +10,11 @@ import {
   SectionContent,
   SectionHeader,
 } from '../../Section'
-import Copy from './Copy.vue'
+import EndpointDetails from './EndpointDetails.vue'
+import EndpointPath from './EndpointPath.vue'
 import ExampleRequest from './ExampleRequest.vue'
 import { ExampleResponses } from './ExampleResponses'
+import TryRequestButton from './TryRequestButton.vue'
 
 defineProps<{
   operation: TransformedOperation
@@ -34,11 +36,20 @@ defineProps<{
               </Anchor>
             </SectionHeader>
           </div>
-          <Copy :operation="operation" />
+          <EndpointDetails :operation="operation" />
         </SectionColumn>
         <SectionColumn>
           <div class="examples">
-            <ExampleRequest :operation="operation" />
+            <ExampleRequest :operation="operation">
+              <template #header>
+                <EndpointPath
+                  class="example-path"
+                  :path="operation.path" />
+              </template>
+              <template #footer>
+                <TryRequestButton :operation="operation" />
+              </template>
+            </ExampleRequest>
             <ExampleResponses
               :operation="operation"
               style="margin-top: 12px" />
@@ -54,8 +65,14 @@ defineProps<{
   position: sticky;
   top: calc(var(--refs-header-height) + 24px);
 }
-
 .deprecated * {
   text-decoration: line-through;
+}
+.example-path {
+  color: var(--theme-color-2, var(--default-theme-color-2));
+  font-family: var(--theme-font-code, var(--default-theme-font-code));
+}
+.example-path :deep(em) {
+  color: var(--theme-color-1, var(--default-theme-color-1));
 }
 </style>