+++ packages/api-reference/src/components/Content/ReferenceTagAccordion.vue
@@ -1,0 +1,39 @@
+<!-- eslint-disable vue/no-unused-properties -->
+<script setup lang="ts">
+import { getTagHash } from '../../hooks'
+import { type Spec, type Tag } from '../../types'
+import { Anchor } from '../Anchor'
+import { SectionContainerAccordion, SectionHeader } from '../Section'
+import MarkdownRenderer from './MarkdownRenderer.vue'
+
+defineProps<{
+  tag: Tag
+  spec: Spec
+  isFirst?: boolean
+}>()
+</script>
+<template>
+  <SectionContainerAccordion class="tag-section">
+    <template #title>
+      <SectionHeader
+        class="tag-name"
+        :level="2">
+        <Anchor :id="getTagHash(tag)">
+          {{ tag.name }}
+        </Anchor>
+      </SectionHeader>
+      <MarkdownRenderer
+        class="tag-description"
+        :value="tag.description" />
+    </template>
+    <slot />
+  </SectionContainerAccordion>
+</template>
+<style scoped>
+.tag-section + .tag-section {
+  margin-top: 20px;
+}
+.tag-name {
+  text-transform: capitalize;
+}
+</style>